Fixed up models

+++ src/storage/sql/__mixin__.py
@@ -1,6 +1,6 @@
-from sqlalchemy import BigInteger
+from sqlalchemy import Integer
 from sqlalchemy.orm import Mapped, mapped_column
 
 
 class IdMixin:
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True)
